pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether-os"
version = "0.1.0"
description = "AETHER real-time multimodal reasoning OS starter monorepo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "numpy>=1.26.0",
  "networkx>=3.3",
  "prometheus-client>=0.20.0",
  "python-json-logger>=2.0.7",
  "redis>=5.0.7",
  "aiokafka>=0.11.0",
  "neo4j>=5.22.0",
  "asyncpg>=0.29.0",
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "orjson>=3.10.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.1",
  "pytest-asyncio>=0.23.7",
  "respx>=0.21.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["packages/aether_core/src", "."]
include = ["aether_core*", "services*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "ASYNC"]

